5.26乌拉乌拉

+++ App/Inc/searchroad.h
@@ -52,8 +52,8 @@ extern void judge_island();
 extern int8 real_L(int8 left);  //用来把左数组的-1变成0
 extern int8 real_R(int8 right); //用来把右数组的-1变成79
 
-extern int16 left_black[CAMERA_H];                          //左边黑线数组
-extern int16 right_black[CAMERA_H];                          //左边黑线数组
-extern int16 middleline[CAMERA_H];  
+extern int8 left_black[CAMERA_H];                          //左边黑线数组
+extern int8 right_black[CAMERA_H];                          //左边黑线数组
+extern int8 middleline[CAMERA_H];  
 
 #endif
